requirements.txt
playwright==1.45.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.134.0
python-dotenv==1.0.0
aiohttp==3.9.5
pydantic==2.7.4